Break up cis2 auth specs
