+++ app/views/users/accounts/team_not_found.html.erb
@@ -1,4 +1,4 @@
-<%=  h1 "Your organisation is not using this service yet" %>
+<%= h1 "Your organisation is not using this service yet" %>
 
 <p>
   The following organisation is not currently set up to use Mavis:
